feat: Add computing G^(2) using the expression (53) with the integral
The variances from zero initial conditions seem to be always positive. This means that the error is likely in the evaluation of the integral (in either analytical expressions or the implementation of these expressions). At this stage, the new results with the integral seem promising, but more testing is needed.