fix: Remove double writing of compartment pointers to the Neuron's list
Currently the length of a dendritic compartment is set to be twice its radius. This is probably not true, but, apparently, .swc files assume all compartments to be spherical in shape, and the spheres can intersect. It seems like in most of the hippocampus morphologies the dendrites start from the inside of the soma, which is represented by 3 spheres of relatively large radius.