fix: Fix an error in all nonstationary covariances computation methods
The error was in setting the 2-nd order matrices and RHSs for nonstationary problem. The factor of 2 should be present only in front of the time derivatives of G^(2)_ij when i=j. Before, the factor 2 was present in all G^(2)_ij, which caused a mismatch with the corresponding results of Gillespie simulations.