doc: Add a full PDE for the generating function to the section on continuous limit
Not finished yet -- work in progress. Thinking of finding continuous limit for spatio-temporal correlations... Need to finish writing for the averages.

Could be better to create a section analogous to Section 2.4.2 "Second-order expansions" for the first order and use these expansions in deriving the continuous limit for the averages (instead of the whole PDE for the generating function).