fix: Fix computing 2nd moments from the system of ODEs written in a matrix form
The error was in computation of "diffusion matrix", i.e. the free term of the ODE for the covariance matrix. In computing the covariance of the noise, I forgot to include terms that come from <\epsilon_i\epsilon_j>, when there is hopping between i and j. In this case, \epsilon_i\epsilon_j = -1 with probability that is O(dt), so cannot be ignored. I need to include this to the notes. NOTE THAT THE CURRENT IMPLEMENTATION IS SUBOPTIMAL -- CONSTRUCTION AND SPANNING OF THE HOPPING_RATE_MATRIX SHOULD BE AVOIDED!