feat: Implement simple version of heterosynaptic plasticity
Currently, neuron is recreated every time the parameters change, and the Analytic_engine then resets the pointer to the new neuron with the altered parameters.

Interesting effect happens when proteins cannot decay in synapses. Changing binding/unbinding rates does not change the average number of proteins in all other compartments of the neuron, but changes the noise.