feat: Make refreshing the neuron for Analytic_engine
To study the effects of changing parameters of the neuron one needs to set new parameters in the corresponding compartment in main() and then refresh the neuron as neuron.refresh(). This keeps expectations and correlations untouched while the parameters of the neuron get altered. Particularly, the hopping rates of junctions are recalculated.

Note that this may be suboptimal when only one parameter changes, as neuron.refresh() resets all the rate constants.